refactor(cdk:utils): move code to src
