+++ jest.config.js
@@ -1,7 +1,7 @@
 module.exports = {
   // u can change this option to a more specific folder for test single component or util when dev
   // for example, ['<rootDir>/packages/components/button']
-  roots: ['<rootDir>/packages'],
+  roots: ['<rootDir>/packages/cdk'],
 
   testEnvironment: 'jsdom',
   transform: {
