+++ packages/cdk/utils/index.ts
@@ -1,7 +1,7 @@
-export * from './convert'
-export * from './dom'
-export * from './logger'
-export * from './propTypes'
-export * from './typeof'
-export * from './uniqueId'
-export * from './vNode'
+export * from './src/convert'
+export * from './src/dom'
+export * from './src/logger'
+export * from './src/propTypes'
+export * from './src/typeof'
+export * from './src/uniqueId'
+export * from './src/vNode'
