+++ packages/cdk/utils/src/vNode.ts
